Sprint 01 03

+++ src/features/posts/routes/handlers/deletePostHandler.ts
@@ -2,14 +2,20 @@ import { Request, Response } from 'express';
 import { HttpStatus } from '../../../../core';
 import { postsRepository } from '../../repositories/posts.repository';
 
-export const deletePostHandler = (req: Request, res: Response) => {
+export async function deletePostHandler(req: Request, res: Response) {
   const id = req.params.id;
-  const post = postsRepository.findById(id);
 
-  if (!post) {
-    res.sendStatus(HttpStatus.NotFound);
+  try {
+    const foundPost = await postsRepository.findById(id);
+
+    if (!foundPost) {
+      res.sendStatus(HttpStatus.NotFound);
+      return;
+    }
+
+    await postsRepository.delete(id);
+    res.sendStatus(HttpStatus.NoContent);
+  } catch (e: unknown) {
+    res.sendStatus(HttpStatus.InternalServerError);
   }
-
-  postsRepository.delete(id);
-  res.sendStatus(HttpStatus.NoContent);
-};
+}
